refactor: improve structure and readability of WebViewBridge and AuthWebViewRequestHandler
